% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_lectures.R
\name{available_lectures}
\alias{available_lectures}
\title{List currently available lectures}
\usage{
available_lectures()
}
\description{
This function lists all currently available lectures.
}
\examples{
available_lectures()

}
