% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_practicals.R
\name{available_practicals}
\alias{available_practicals}
\title{List currently available practicals}
\usage{
available_practicals(verbose = FALSE)
}
\arguments{
\item{verbose}{Add the description text if TRUE (default FALSE)}
}
\description{
This function lists all currently available lectures.
}
\examples{
available_lectures()

}
