% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_project.R
\name{dummy_project}
\alias{dummy_project}
\title{dummy_project}
\usage{
dummy_project()
}
\description{
This function will create an RStudio project called `project0201` and
populate it with `0105-run-birth-death.R` and `0105-step-birth-death.R`.
}
