% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_materials.R
\name{export_materials}
\alias{export_materials}
\title{export_materials}
\usage{
export_materials(materials, type = "exercises")
}
\arguments{
\item{materials}{materials}

\item{type}{content type}
}
\description{
export_materials
}
\keyword{internal}
