% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate_package.R
\name{populate_package}
\alias{populate_package}
\title{populate_package}
\usage{
populate_package(force = FALSE)
}
\arguments{
\item{force}{- overwrite existing files if TRUE (default FALSE)}
}
\description{
This function will create an RStudio project and populate it with
`0105-run-birth-death.R`, and `0105-step-birth-death.R`.
}
