% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practical_series_one.R
\name{practical_series_one}
\alias{practical_series_one}
\title{practical_series_one}
\usage{
practical_series_one(force = FALSE)
}
\arguments{
\item{force}{- overwrite existing files if TRUE (default FALSE)}
}
\description{
This function will populate the current working directory with scripts and
functions from practical series 1.
}
