% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_analog}
\alias{read_analog}
\title{Read an analog value fromt the Raspberry Pi}
\usage{
read_analog(chan)
}
\arguments{
\item{chan}{An integer vector of channel(s) for which we should
read the analog value.}
}
\value{
An integer vector describing the current analog value on the
   specified channel(s).
}
\description{
Read an analog value fromt the Raspberry Pi
}

