% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_poll}
\alias{read_poll}
\title{Reads From a Poll}
\usage{
read_poll(chan)
}
\arguments{
\item{chan}{The channel for which you want the unread values.}
}
\description{
Read any unread values from an active poll.
}
\details{
If more values have accumulated than the buffer had room for, this function
will throw a warning and you will receive the most recent data.
}

