% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lecture.R
\name{run_lecture}
\alias{run_lecture}
\title{Run lecture}
\usage{
run_lecture(name)
}
\arguments{
\item{name}{Name of the lecture}
}
\description{
This function runs a lecture from the RPiR course.
}
\examples{
\dontrun{
run_lecture("RPiR1-intro")
}

}
