% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_practical.R
\name{run_practical}
\alias{run_practical}
\title{Run practical}
\usage{
run_practical(name)
}
\arguments{
\item{name}{Name of the practical}
}
\description{
This function runs a lecture from the RPiR course.
}
\examples{
\dontrun{
run_practical("practical1-1")
}

}
