% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_no_globals.R
\name{assert_no_globals}
\alias{assert_no_globals}
\title{assert_no_globals}
\usage{
assert_no_globals(test_function, name = deparse1(substitute(test_function)))
}
\arguments{
\item{test_function}{function to check for global variables}

\item{name}{optionally (otherwise taken from test_function) name of function}
}
\value{
nothing
}
\description{
This function will stop code execution with an error if the function
passed in has a global variable
}
