% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_no_globals.R
\name{assert_no_globals}
\alias{assert_no_globals}
\title{assert_no_globals}
\usage{
assert_no_globals(test_function, name = deparse1(substitute(test_function)))
}
\arguments{
\item{test_function}{function to check for global variables}

\item{name}{(optional) the name of function as a string -- passed to the
error message}
}
\value{
nothing
}
\description{
This function will stop code execution with an error if the function
passed in has a global variable
}
\examples{
\dontrun{
broken_function <- function() print(y)
assert_no_globals(broken_function)
assert_no_globals(broken_function, name = "the_name_of_my_function")
}

}
