% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_lectures.R
\name{available_lectures}
\alias{available_lectures}
\title{available_lectures}
\usage{
available_lectures(package)
}
\arguments{
\item{package}{Package name, *i.e. * "RPiR"}
}
\description{
available_lectures
}
\examples{
\dontrun{
available_lectures("RPiR")
}

}
