% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_package.R
\name{dummy_package}
\alias{dummy_package}
\title{dummy_package}
\usage{
dummy_package(package_name)
}
\arguments{
\item{package_name}{a \code{string} corresponding to the name you want to
give your package; this should be your GitHub user name.

 @keywords internal}
}
\description{
This function will create an RStudio project and
populate it with `0105-run-birth-death.R` (now `run-birth-death.R`), and
`0105-step-birth-death.R` (now `step_deterministic_birth_death.R`).
}
